Move EmptyCart and CheckoutCompleteRedirecting inside WPCheckout
