+++ client/my-sites/checkout/composite-checkout/composite-checkout.tsx
@@ -1,18 +1,8 @@
 import { useStripe } from '@automattic/calypso-stripe';
 import colorStudio from '@automattic/color-studio';
-import {
-	CheckoutProvider,
-	CheckoutStepAreaWrapper,
-	MainContentWrapper,
-	SubmitButtonWrapper,
-	checkoutTheme,
-	Button,
-	useTransactionStatus,
-	TransactionStatus,
-} from '@automattic/composite-checkout';
+import { CheckoutProvider, checkoutTheme } from '@automattic/composite-checkout';
 import { useShoppingCart } from '@automattic/shopping-cart';
 import { useIsWebPayAvailable, isValueTruthy } from '@automattic/wpcom-checkout';
-import { ThemeProvider } from '@emotion/react';
 import { useSelect } from '@wordpress/data';
 import debugFactory from 'debug';
 import { useTranslate } from 'i18n-calypso';
@@ -29,8 +19,6 @@ import { recordAddEvent } from 'calypso/lib/analytics/cart';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import wp from 'calypso/lib/wp';
 import useSiteDomains from 'calypso/my-sites/checkout/composite-checkout/hooks/use-site-domains';
-import useValidCheckoutBackUrl from 'calypso/my-sites/checkout/composite-checkout/hooks/use-valid-checkout-back-url';
-import { leaveCheckout } from 'calypso/my-sites/checkout/composite-checkout/lib/leave-checkout';
 import {
 	translateCheckoutPaymentMethodToWpcomPaymentMethod,
 	translateCheckoutPaymentMethodToTracksPaymentMethod,
@@ -40,12 +28,9 @@ import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { updateContactDetailsCache } from 'calypso/state/domains/management/actions';
 import { errorNotice, infoNotice } from 'calypso/state/notices/actions';
 import getIsIntroOfferRequesting from 'calypso/state/selectors/get-is-requesting-into-offers';
-import getPreviousRoute from 'calypso/state/selectors/get-previous-route';
 import isPrivateSite from 'calypso/state/selectors/is-private-site';
 import isAtomicSite from 'calypso/state/selectors/is-site-automated-transfer';
 import { isJetpackSite } from 'calypso/state/sites/selectors';
-import { CheckoutCompleteRedirecting } from './components/checkout-complete-redirecting';
-import EmptyCart from './components/empty-cart';
 import WPCheckout from './components/wp-checkout';
 import useActOnceOnStrings from './hooks/use-act-once-on-strings';
 import useAddProductsFromUrl from './hooks/use-add-products-from-url';
@@ -79,7 +64,6 @@ import { StoredCard } from './types/stored-cards';
 import { emptyManagedContactDetails } from './types/wpcom-store-state';
 import type { PaymentProcessorOptions } from './types/payment-processors';
 import type { CheckoutPageErrorCallback } from '@automattic/composite-checkout';
-import type { ResponseCart } from '@automattic/shopping-cart';
 import type {
 	ManagedContactDetails,
 	CountryListItem,
@@ -138,7 +122,6 @@ export default function CompositeCheckout( {
 	isUserComingFromLoginForm?: boolean;
 	customizedPreviousPath?: string;
 } ): JSX.Element {
-	const previousPath = useSelector( getPreviousRoute );
 	const translate = useTranslate();
 	const isJetpackNotAtomic =
 		useSelector(
@@ -318,9 +301,10 @@ export default function CompositeCheckout( {
 		);
 	} );
 
-	const errors = responseCart.messages?.errors ?? [];
+	const responseCartErrors = responseCart.messages?.errors ?? [];
 	const areThereErrors =
-		[ ...errors, cartLoadingError, cartProductPrepError ].filter( isValueTruthy ).length > 0;
+		[ ...responseCartErrors, cartLoadingError, cartProductPrepError ].filter( isValueTruthy )
+			.length > 0;
 
 	const {
 		isRemovingProductFromCart,
@@ -681,64 +665,6 @@ export default function CompositeCheckout( {
 		reduxDispatch( infoNotice( translate( 'Redirecting to payment partner…' ) ) );
 	}, [ reduxDispatch, translate ] );
 
-	const jetpackCheckoutBackUrl = useValidCheckoutBackUrl( updatedSiteSlug );
-
-	const goToPreviousPage = () =>
-		leaveCheckout( {
-			siteSlug,
-			jetpackCheckoutBackUrl,
-			previousPath: customizedPreviousPath || previousPath,
-			tracksEvent: 'calypso_checkout_composite_empty_cart_clicked',
-		} );
-
-	const { transactionStatus } = useTransactionStatus();
-
-	if ( transactionStatus === TransactionStatus.COMPLETE ) {
-		debug( 'rendering post-checkout redirecting page' );
-		return (
-			<Fragment>
-				<PageViewTracker path={ analyticsPath } title="Checkout" properties={ analyticsProps } />
-				<ThemeProvider theme={ theme }>
-					<MainContentWrapper>
-						<CheckoutStepAreaWrapper>
-							<CheckoutCompleteRedirecting />
-						</CheckoutStepAreaWrapper>
-					</MainContentWrapper>
-				</ThemeProvider>
-			</Fragment>
-		);
-	}
-
-	if (
-		shouldShowEmptyCartPage( {
-			responseCart,
-			areWeRedirecting: isRemovingProductFromCart,
-			areThereErrors,
-			isCartPendingUpdate,
-			isInitialCartLoading,
-		} )
-	) {
-		debug( 'rendering empty cart page' );
-
-		return (
-			<Fragment>
-				<PageViewTracker path={ analyticsPath } title="Checkout" properties={ analyticsProps } />
-				<ThemeProvider theme={ theme }>
-					<MainContentWrapper>
-						<CheckoutStepAreaWrapper>
-							<EmptyCart />
-							<SubmitButtonWrapper>
-								<Button buttonType="primary" fullWidth onClick={ goToPreviousPage }>
-									{ translate( 'Go back' ) }
-								</Button>
-							</SubmitButtonWrapper>
-						</CheckoutStepAreaWrapper>
-					</MainContentWrapper>
-				</ThemeProvider>
-			</Fragment>
-		);
-	}
-
 	return (
 		<Fragment>
 			<QueryIntroOffers siteId={ updatedSiteId } />
@@ -771,6 +697,10 @@ export default function CompositeCheckout( {
 				initiallySelectedPaymentMethodId={ paymentMethods?.length ? paymentMethods[ 0 ].id : null }
 			>
 				<WPCheckout
+					customizedPreviousPath={ customizedPreviousPath }
+					isRemovingProductFromCart={ isRemovingProductFromCart }
+					areThereErrors={ areThereErrors }
+					isInitialCartLoading={ isInitialCartLoading }
 					addItemToCart={ addItemAndLog }
 					changePlanLength={ changePlanLength }
 					countriesList={ countriesList }
@@ -837,35 +767,4 @@ function getAnalyticsPath(
 	}
 
 	return { analyticsPath, analyticsProps };
-}
-
-function shouldShowEmptyCartPage( {
-	responseCart,
-	areWeRedirecting,
-	areThereErrors,
-	isCartPendingUpdate,
-	isInitialCartLoading,
-}: {
-	responseCart: ResponseCart;
-	areWeRedirecting: boolean;
-	areThereErrors: boolean;
-	isCartPendingUpdate: boolean;
-	isInitialCartLoading: boolean;
-} ): boolean {
-	if ( responseCart.products.length > 0 ) {
-		return false;
-	}
-	if ( areWeRedirecting ) {
-		return false;
-	}
-	if ( areThereErrors ) {
-		return true;
-	}
-	if ( isCartPendingUpdate ) {
-		return false;
-	}
-	if ( isInitialCartLoading ) {
-		return false;
-	}
-	return true;
 }
